added repo

+++ app/src/main/java/com/forntoh/twofoureight/ui/components/Footer.kt
@@ -22,7 +22,7 @@ import java.util.concurrent.TimeUnit
 
 @Composable
 fun Footer(
-    moves: Long,
+    moves: Int,
     millisecondsElapsed: Long,
     modifier: Modifier = Modifier
 ) {
